# src/QbAutopost.Api/wwwroot/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QbAutopost API",
    "version": "1.0",
    "description": "Turns a job folder or a batch of JSON rows into posted QuickBooks Desktop transactions.\n\nThis document is **hand-authored** (`src/QbAutopost.Api/wwwroot/openapi.json`): .NET 8 has no built-in OpenAPI generator and Swashbuckle is excluded, so nothing generates it from the code. It is kept honest by `OpenApiDriftTests`, which walks the application's own route table and fails the build when a route is undocumented, an operation has no route, or an operation is missing its summary, its response schema or the scope it needs. `x-required-scope` on each operation is the scope the middleware actually enforces, not a description of it.\n\nThe reference page and this document are served only when `Api:Reference:Enabled` is true, and both need the `health:read` scope."
  },
  "servers": [
    { "url": "/", "description": "This installation" }
  ],
  "security": [
    { "apiKeyAuth": [] }
  ],
...
  "paths": {
...
    "/api/v1/quickbooks/connection/test": {
      "post": {
        "tags": ["QuickBooks"],
        "operationId": "testConnection",
        "summary": "A full session round trip, timed step by step.",
        "description": "Its own budget (`QuickBooks:ConnectionTestTimeoutSeconds`, default 180 s), deliberately longer than the busy timeout: on the owner's server a certificate dialog held the first session 101 s, and a slow answer is not a broken one. Changes nothing.",
        "x-required-scope": "qb:read",
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConnectionTestRequest" } } }
        },
        "responses": {
          "200": {
            "description": "The round trip finished; `ok` and `steps` say how it went.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConnectionTestResult" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
    "/api/v1/quickbooks/company-file/validate": {
      "post": {
        "tags": ["QuickBooks"],
        "operationId": "validateCompanyFile",
        "summary": "Is this company file the configured one, open in QuickBooks, and backed up?",
        "description": "Read-only: no qbXML is sent. A *different* company file open in QuickBooks is an error, not a warning — it is the one misconfiguration that posts real money into the wrong ledger.",
        "x-required-scope": "qb:read",
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyFileRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Every error-severity check passed.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyFileValidation" } } }
          },
          "422": {
            "description": "A check of severity `error` failed; the body lists every check.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompanyFileValidation" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
    "/api/v1/quickbooks/transactions/validate": {
      "post": {
        "tags": ["QuickBooks"],
        "operationId": "validateTransactions",
        "summary": "Would these rows post? The same reader, mapper and gates, stopping short of the SDK call.",
        "description": "Offline and never writes, so a batch can be checked while the QuickBooks server is shut down. 400 when the request contradicts itself (a control total that does not add up, no rows, too many); 422 when it was understood and a gate held a row.",
        "x-required-scope": "qb:read",
        "parameters": [
          {
            "name": "includeQbXml",
            "in": "query",
            "required": false,
            "description": "Return the qbXML request body as well. It carries every amount and name in the batch, so it is served only to a key holding `qb:debug`; a caller without it is told so rather than left wondering.",
            "schema": { "type": "boolean", "default": false }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Every row would post.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationResult" } } }
          },
          "422": {
            "description": "The request was understood and at least one row was held or skipped.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationResult" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/quickbooks/transactions": {
      "post": {
        "tags": ["QuickBooks"],
        "operationId": "postTransactions",
        "summary": "Post JSON rows to QuickBooks and return the batch result.",
        "description": "The only route that sends transactions to QuickBooks. The batch is on disk before it is queued, so there is always something to poll. `dryRun` (default `DryRunDefault`) ends in `ready` with nothing sent. If the work outlasts `Api:SyncPostTimeoutSeconds` the answer is 202 with a batch to poll and the work carries on; `Prefer: respond-async` asks for that immediately.",
        "x-required-scope": "qb:post",
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          {
            "name": "Prefer",
            "in": "header",
            "required": false,
            "description": "`respond-async` to be answered 202 without waiting.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DirectRequest" } } }
        },
        "responses": {
          "200": {
            "description": "The batch finished (or was a dry run).",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BatchResult" } } }
          },
          "202": {
            "description": "Still posting; poll `location`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BatchAccepted" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
    "/api/v1/quickbooks/lists": {
      "get": {
        "tags": ["QuickBooks"],
        "operationId": "getLists",
        "summary": "The cached QuickBooks account, vendor and customer names.",
        "description": "Read from `qb-lists.json`, so it answers while QuickBooks is shut down and never queues behind a post. A `syncedUtc` of `null` with empty lists means the sync has not run yet; it is not an error.",
        "x-required-scope": "qb:read",
        "responses": {
          "200": {
            "description": "The names as the last sync left them.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QbLists" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/quickbooks/lists/sync": {
      "post": {
        "tags": ["QuickBooks"],
        "operationId": "syncLists",
        "summary": "Read the QuickBooks account, vendor and customer lists into qb-lists.json.",
        "description": "One read-only message set (active records only). It runs on the single job worker, so it cannot overlap a job or a post. This is the v1 spelling; `POST /api/v1/qb/sync-lists` is the same handler under the pre-M9 name and will go when the legacy routes do (Q-53, Q-69).",
        "x-required-scope": "qb:read",
        "responses": {
          "200": {
            "description": "The lists were read and written.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListSyncResult" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "502": { "$ref": "#/components/responses/QuickBooksError" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
    "/api/v1/qb/sync-lists": {
      "post": {
        "tags": ["QuickBooks"],
        "operationId": "syncListsFlat",
        "summary": "Deprecated spelling of POST /api/v1/quickbooks/lists/sync; same handler, same body.",
        "description": "Kept because callers written before M9 use it. Move to `/api/v1/quickbooks/lists/sync`; this path goes when `Api:LegacyRoutes` is turned off (Q-53).",
        "deprecated": true,
        "x-required-scope": "qb:read",
        "responses": {
          "200": {
            "description": "The lists were read and written.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListSyncResult" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "502": { "$ref": "#/components/responses/QuickBooksError" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "`X-Api-Key: <key>` — the header this build reads, and the one every operation is secured with here. Each key belongs to one caller in `clients.json` and carries only the scopes that caller was given; `x-required-scope` on every operation names the one it needs."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer <key>`. api-v1 §2.2 calls this the preferred form and FR-A-18 asks this document to declare it — but **this build does not accept it yet**: `ApiKeyMiddleware` reads only `X-Api-Key`. It is declared, not offered, so no reader is sent down a road the app refuses (Q-67)."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Replaying a request under the same key returns the stored answer and sends nothing to QuickBooks a second time; the same key with a different body is a 409.",
        "schema": { "type": "string", "maxLength": 200 }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed or self-inconsistent request. Nothing was posted.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "Unauthorized": {
        "description": "No key, or a key no caller holds.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "Forbidden": {
        "description": "The key is known but lacks the scope this operation needs.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
...
      "Conflict": {
        "description": "A state conflict, or an `Idempotency-Key` replayed with a different body.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "TooManyRequests": {
        "description": "Rate limited; `Retry-After` says for how long.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
...
      "Unavailable": {
        "description": "QuickBooks or Hermes is unreachable or busy, or the app is not ready.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      }
    },
    "schemas": {
...
      "ConnectionTestRequest": {
        "type": "object",
        "properties": {
          "companyFile": {
            "type": "string",
            "nullable": true,
            "description": "Refused unless `QuickBooks:AllowCompanyFileOverride` is on and the path is inside `QuickBooks:AllowedCompanyFolders`."
          },
          "timeoutSeconds": { "type": "integer", "format": "int32", "nullable": true },
          "includeCompanyInfo": { "type": "boolean", "nullable": true }
        }
      },
      "ConnectionTestResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "companyFile": { "type": "string", "nullable": true },
          "product": { "type": "string", "nullable": true },
          "qbXmlVersion": { "type": "string" },
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "enum": ["waitForGateway", "hostQuery", "companyFile"] },
                "ok": { "type": "boolean" },
                "ms": { "type": "integer", "format": "int64" },
                "message": { "type": "string", "nullable": true }
              }
            }
          },
          "totalMs": { "type": "integer", "format": "int64" },
          "message": { "type": "string" }
        }
      },
      "CompanyFileRequest": {
        "type": "object",
        "properties": {
          "companyFile": { "type": "string", "nullable": true },
          "requireBackup": { "type": "boolean", "nullable": true, "default": false, "description": "Ask for a guarantee rather than a warning (T-917). `true` raises the `backupFreshness` check from warning to error, so a stale backup — or a server with no `QuickBooks:BackupFolder` configured, which cannot confirm one — makes `ok` false and the response 422. What the flag should mean was never specified; see Q-72." }
        }
      },
      "CompanyFileValidation": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "companyFile": { "type": "string", "nullable": true },
          "checks": { "type": "array", "items": { "$ref": "#/components/schemas/Check" } },
          "message": { "type": "string" }
        }
      },
      "DirectRequest": {
        "type": "object",
        "required": ["transactions"],
        "properties": {
          "reference": { "type": "string", "nullable": true, "description": "The caller's own name for this batch." },
          "dryRun": {
            "type": "boolean",
            "nullable": true,
            "description": "Omitted follows `DryRunDefault`, which ships true."
          },
          "controlTotal": {
            "type": "number",
            "format": "double",
            "nullable": true,
            "description": "The caller's own sum over every submitted row. It is checked, never corrected: a mismatch refuses the whole batch and no amount is adjusted to make it agree."
          },
          "allowModelAccounts": {
            "type": "boolean",
            "default": false,
            "description": "NOT IMPLEMENTED (T-915). Sending `true` is refused with 400: tiers 3-4 never run on the direct path, so the flag would silently do nothing. Give each row a `lineAccount`, or add a rule for its last four. Whether this is ever implemented is open question Q-51; the `qb:post:ai` scope is reserved for it and is checked nowhere today."
          },
          "transactions": { "type": "array", "items": { "$ref": "#/components/schemas/DirectRow" } }
        }
      },
...
      "ValidationResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "counts": {
            "type": "object",
            "description": "`wouldPost + held + skipped` is `submitted`, so the arithmetic of the answer can be checked too. `duplicates` is part of `skipped`.",
            "properties": {
              "submitted": { "type": "integer", "format": "int32" },
              "wouldPost": { "type": "integer", "format": "int32" },
              "held": { "type": "integer", "format": "int32" },
              "skipped": { "type": "integer", "format": "int32" },
              "duplicates": { "type": "integer", "format": "int32" }
            }
          },
          "totals": {
            "type": "object",
            "properties": {
              "submitted": { "type": "number", "format": "double" },
              "wouldPost": { "type": "number", "format": "double" },
              "controlTotal": { "type": "number", "format": "double", "nullable": true },
              "controlTotalOk": { "type": "boolean" }
            }
          },
          "rows": { "type": "array", "items": { "$ref": "#/components/schemas/PlanRow" } },
          "unknownNames": {
            "type": "object",
            "description": "Names in the batch that `qb-lists.json` does not have.",
            "properties": {
              "accounts": { "type": "array", "items": { "type": "string" } },
              "payees": { "type": "array", "items": { "type": "string" } }
            }
          },
          "qbXml": {
            "type": "object",
            "properties": {
              "requestCount": { "type": "integer", "format": "int32" },
              "byType": { "type": "object", "additionalProperties": { "type": "integer", "format": "int32" } },
              "request": { "type": "string", "nullable": true, "description": "Only with `includeQbXml=true` and the `qb:debug` scope." },
              "note": { "type": "string", "nullable": true }
            }
          }
        }
      },
...
      "BatchAccepted": {
        "type": "object",
        "properties": {
          "batchId": { "type": "string" },
          "status": { "$ref": "#/components/schemas/JobStatus" },
          "location": { "type": "string", "description": "Poll this for the finished batch." }
        }
      },
...
    }
  }
}